+++ src/main/java/br/com/abce/sge/entity/AvaliacaoItemEntityPK.java
@@ -6,26 +6,26 @@ import java.io.Serializable;
 import java.util.Objects;
 
 public class AvaliacaoItemEntityPK implements Serializable {
-    private int avaliacaoId;
-    private int avaliacaoMotivoId;
+    private long avaliacaoId;
+    private long avaliacaoMotivoId;
 
     @Column(name = "avaliacao_id")
     @Id
-    public int getAvaliacaoId() {
+    public long getAvaliacaoId() {
         return avaliacaoId;
     }
 
-    public void setAvaliacaoId(int avaliacaoId) {
+    public void setAvaliacaoId(long avaliacaoId) {
         this.avaliacaoId = avaliacaoId;
     }
 
     @Column(name = "avaliacao_motivo_id")
     @Id
-    public int getAvaliacaoMotivoId() {
+    public long getAvaliacaoMotivoId() {
         return avaliacaoMotivoId;
     }
 
-    public void setAvaliacaoMotivoId(int avaliacaoMotivoId) {
+    public void setAvaliacaoMotivoId(long avaliacaoMotivoId) {
         this.avaliacaoMotivoId = avaliacaoMotivoId;
     }
